test(Node execution): Move tests into own file
